Move ByteUtil to infrastructure:unsigned.
