+++ infrastructure/unsigned/src/main/java/tech/pegasys/teku/infrastructure/unsigned/ByteUtil.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.bytes;
+package tech.pegasys.teku.infrastructure.unsigned;
 
 import static com.google.common.base.Preconditions.checkArgument;
