+++ infrastructure/unsigned/src/test/java/tech/pegasys/teku/infrastructure/unsigned/ByteUtilTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.bytes;
+package tech.pegasys.teku.infrastructure.unsigned;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
