+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/rpc/core/RpcResponseDecoder.java
@@ -13,9 +13,9 @@
 
 package tech.pegasys.teku.networking.eth2.rpc.core;
 
+import static tech.pegasys.teku.infrastructure.unsigned.ByteUtil.toByteExactUnsigned;
+import static tech.pegasys.teku.infrastructure.unsigned.ByteUtil.toUnsignedInt;
 import static tech.pegasys.teku.networking.eth2.rpc.core.RpcResponseStatus.SUCCESS_RESPONSE_CODE;
-import static tech.pegasys.teku.util.bytes.ByteUtil.toByteExactUnsigned;
-import static tech.pegasys.teku.util.bytes.ByteUtil.toUnsignedInt;
 
 import io.netty.buffer.ByteBuf;
 import java.util.ArrayList;
